+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   - Fix caching
   - Now it's working fine with the `vybe-games` builds
   - Create `vc/eval*` to debug standalone forms
+  - Accept meta `^:void` in `if`/`cond` expressions to signal to the C compiler that we don't care about the return
 - Add `vf/defsystem-c`
 
 ## v0.7.469